Moving run method to runnable module
